Fall back to SHA-1 for base64/binary hashCode
MD5 digest algorithm is unavailable on a FIPS JVM, so fall back to SHA-1
for generating object hashCodes in this case.  Only the first 32 bits
are used, so the length difference will not matter.
